+++ test/lyapunovs/discrete_lyapunov.jl
@@ -69,5 +69,9 @@ end
     @test 0.418 < λ1[1] < 0.422
     @test 0.418 < λ2[1] < 0.422
     @test isapprox(λ1, λ2; rtol = 1e-3)
+
+    ls, ts = lyapunov(ds1, 1000000, Val{true})
+    @test 0.418 < λs[end] < 0.422
+    @test length(ts) == length(ls)
   end
 end